feat: updating to lastes & adding great optimzations
