+++ package.json
@@ -11,17 +11,19 @@
 		"lint": "next lint"
 	},
 	"dependencies": {
-		"next": "11.0.0",
-		"next-translate": "^1.0.7",
+		"critters": "^0.0.14",
+		"next": "^12.0.4",
+		"next-translate": "^1.2.0",
+		"preact": "^10.5.15",
 		"react": "17.0.2",
 		"react-dom": "17.0.2"
 	},
 	"devDependencies": {
-		"@types/react": "17.0.11",
-		"eslint": "^7.29.0",
-		"eslint-config-next": "^11.0.0",
-		"next-sitemap": "^1.6.116",
-		"sass": "^1.35.1",
-		"typescript": "4.3.4"
+		"@types/react": "17.0.36",
+		"eslint": "^7.32.0",
+		"eslint-config-next": "^12.0.4",
+		"next-sitemap": "^1.6.203",
+		"sass": "^1.43.4",
+		"typescript": "^4.5.2"
 	}
 }
