+++ pages/_document.tsx
@@ -65,17 +65,6 @@ class MyDocument extends Document {
 						href="https://fonts.googleapis.com/css2?family=Comfortaa:wght@400;500;700&family=Tajawal:wght@400;500;700&display=swap"
 						rel="stylesheet"/>
 
-					{/* Google Analytics */}
-					<script async src="https://www.googletagmanager.com/gtag/js?id=G-H09VYEZ3MD"/>
-					<script dangerouslySetInnerHTML={{
-						__html: `
-						 window.dataLayer = window.dataLayer || [];
-						 function gtag(){dataLayer.push(arguments);}
-						 gtag('js', new Date());
-						 gtag('config', 'G-H09VYEZ3MD');
-					`
-					}}/>
-
 					{/* Structured Data */}
 					<script
 						type="application/ld+json"
